Fix secp256k1_ecdsa_signature_normalize
This function was untested, and was found to be
broken when I was using it in tests.

This allows an arbitrary value to be passed as
the second parameter, where the normalized signature
resource will be written.

It also fixes a logic error I had in the function,
which meant the signature would only be written IFF
the sigin was not in normalized form.