Expose signature nonce generation functions
These are small and easy to miss, so I never thought to
do it before. However, PHP doesn't allow you to pass NULL for a callable
argument - all that is possible is not to provide the argument.

The secp256k1_ecdsa_sign function and bipschnorr version are more
flexible however, and allow passing NULL for the noncefp, while passing
some data in the next argument.

The only way to be able to pass extra data and use the default function
from PHP was pass the function by name, hence we had to expose them.