Build system: allow customizing modules
Previously, we built the extension assuming all modules were available.
For most installations however, only the core library is required, and
not the experimental modules.

The default behavior remains to build with support for all modules, however
if the new build option `--with-secp256k1-config` is set, all modules will
be disabled unless specifically enabled by the following flags:

 - --with-module-recovery
 - --with-module-ecdh
 - --with-module-schnorrsig